pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-jose>=3.3.0",
    "pymongo>=4.8.0",
    "graphql-core>=3.2.3",
    "pytest>=8.3.2",
    "python-dotenv>=1.0.1",
    "passlib>=1.7.4",
    "bcrypt>=4.2.0",
    "pydantic-settings>=2.4.0",
    "starlette>=0.38.2",
    "email-validator>=2.2.0",
    "transformers>=4.44.2",
    "pip>=24.2",
    "torch",
    "mlx>=0.17.2",
    "sentencepiece>=0.2.0",
    "mlx-lm>=0.18.1",
    "strawberry-graphql>=0.246.1",
    "fastapi>=0.115.0",
    "fastapi-cli>=0.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"
